+++ api/core/v1alpha1/crds/druid.gardener.cloud_etcds_without_cel.yaml
