+++ test/it/crdvalidation/etcd/validation_test.go
@@ -6,10 +6,10 @@ package etcd
 
 import (
 	"fmt"
+	"github.com/gardener/etcd-druid/api/core/v1alpha1/crds"
 	"os"
 	"testing"
 
-	"github.com/gardener/etcd-druid/api/core/crds"
 	"github.com/gardener/etcd-druid/test/it/assets"
 	"github.com/gardener/etcd-druid/test/it/setup"
 )
